backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djoser==2.2.2
Pillow==10.1.0
drf-yasg==1.21.7